+++ cmd/root.go
@@ -1,11 +1,10 @@
 package cmd
 
 import (
+	"github.com/resonatehq/resonate/internal/creds"
 	"log/slog"
 	"os"
 	"strings"
-	"errors"
-	"github.com/resonatehq/resonate/internal/app/subsystems/api/http"
 
 	"github.com/resonatehq/resonate/cmd/dst"
 	"github.com/resonatehq/resonate/cmd/promises"
@@ -50,18 +49,6 @@ func init() {
 	rootCmd.SetErr(os.Stderr)
 }
 
-
-var CredsFromFile http.CredentialsList
-
-func GetCrednetials() (http.CredentialsList, error) {
-	if len(CredsFromFile.Users) == 0 {
-		slog.Error("CredentialsList is empty", "error", errors.New("Credentials are empty."))
-		return http.CredentialsList{}, errors.New("Credentials are empty.")
-	} else {
-		return CredsFromFile, nil
-	}
-}
-
 func initCreds() {
 	if credFile != "" {
 		viper.SetConfigFile(credFile)
@@ -77,7 +64,7 @@ func initCreds() {
 		return
 	}
 
-	err = viper.Unmarshal(&CredsFromFile)
+	err = viper.Unmarshal(&creds.CredsFromFile)
 	if err != nil {
 		slog.Error("Unable to decode creds from file into struct", "error", err)
 		return
@@ -87,7 +74,7 @@ func initCreds() {
 func initConfig() {
 	initCreds()
 	viper.Reset()
-	
+
 	if cfgFile != "" {
 		viper.SetConfigFile(cfgFile)
 	} else {
